+++ db/pvault.sql
@@ -329,7 +329,8 @@ CREATE TABLE `settings` (
   `pv_scale_enabled` INT NOT NULL DEFAULT '0',
   `pv_host` VARCHAR(255) NOT NULL DEFAULT 'localhost',
   `sds_disclaimer` MEDIUMTEXT NOT NULL DEFAULT 'PLEASE ADD A PROPER DISCLAIMER MESSAGE',
-  `bs_theme` VARCHAR(255) NOT NULL DEFAULT 'light'
+  `bs_theme` VARCHAR(255) NOT NULL DEFAULT 'light',
+  `temp_sys` VARCHAR(255) NOT NULL DEFAULT '°C'
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
 
 INSERT INTO `settings` (`id`, `currency`, `top_n`, `heart_n`, `base_n`, `chem_vs_brand`, `grp_formula`, `brandName`, `brandAddress`, `brandEmail`, `brandPhone`, `brandLogo`) VALUES