+++ db/updates/update_11.3-11.4.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `settings` ADD `temp_sys` VARCHAR(255) NOT NULL DEFAULT '°C' AFTER `bs_theme`;