+++ releasenotes.md
@@ -1,5 +1,6 @@
 Whats New in v11.4
 --------------------------
+- Add temprature measurement unit
 - Add solubility options instead of a free text for ingredients
 - Cleaner and better view for formula comparison
 - This release may include more changes, for full details please refer to the CHANGELOG